upgraded some things
