+++ public/style.css
@@ -186,6 +186,12 @@ td.column-td {
   font-size: 1.5rem;
 }
 
+.jaam-icon {
+  width: 3rem;
+  height: 2rem;
+  margin: 1rem;
+}
+
 .icon-small {
   font-size: 1rem;
   margin-left: 0.5rem;
